src/face-cut/requirements.txt
pillow
boto3~=1.35.99